Fix bugs to get it working on hardware again

+++ src/math/mathf.h
@@ -16,7 +16,7 @@ float sqrtf(float in);
 
 float cosf(float in);
 float sinf(float in);
-float fabs(float in);
+float fabsf(float in);
 float floorf(float in);
 
 #endif
